requirements.txt
langchain>=0.1.16
langchain-community>=0.0.36
langchain-google-genai>=0.0.9
langchain-core>=0.1.50
huggingface_hub>=0.22.2
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
tavily-python>=0.3.2   